src/main/java/io/gdfbarbosa/algorithms/design/HitCounter.java
package io.gdfbarbosa.algorithms.design;

import java.util.LinkedList;
import java.util.Queue;

/**
 * <a href="https://leetcode.com/problems/design-hit-counter/">362. Design Hit Counter</a>
 */
public class HitCounter {
    private final Queue<Integer> hits;

    /** Initialize your data structure here. */
    public HitCounter() {
        this.hits = new LinkedList<>();
    }

    /** Record a hit.
     @param timestamp - The current timestamp (in seconds granularity). */
    public void hit(int timestamp) {
        this.hits.add(timestamp);
    }

    /** Return the number of hits in the past 5 minutes.
     @param timestamp - The current timestamp (in seconds granularity). */
    public int getHits(int timestamp) {
        while (!this.hits.isEmpty()) {
            int diff = timestamp - this.hits.peek();
            if (diff >= 300) this.hits.remove();
            else break;
        }
        return this.hits.size();
    }
}
